src/requirements.txt
python-docx
qdrant_client
streamlit